+++ src/main/java/com/commercehub/gradle/plugin/avro/Constants.java
@@ -3,8 +3,6 @@ package com.commercehub.gradle.plugin.avro;
 import org.apache.avro.compiler.specific.SpecificCompiler.FieldVisibility;
 import org.apache.avro.generic.GenericData.StringType;
 
-import static java.lang.Boolean.TRUE;
-
 /**
  * Various constants needed by the plugin.
  *
@@ -18,8 +16,8 @@ class Constants {
     static final String DEFAULT_STRING_TYPE = StringType.String.name();
     static final String DEFAULT_FIELD_VISIBILITY = FieldVisibility.PUBLIC_DEPRECATED.name();
     static final String DEFAULT_TEMPLATE_DIR = System.getProperty("org.apache.avro.specific.templates",
-                         "/org/apache/avro/compiler/specific/templates/java/classic/");
-    static final Boolean DEFAULT_CREATE_SETTERS = TRUE;
+                                                                  "/org/apache/avro/compiler/specific/templates/java/classic/");
+    static final boolean DEFAULT_CREATE_SETTERS = true;
 
     static final String SCHEMA_EXTENSION = "avsc";
     static final String PROTOCOL_EXTENSION = "avpr";
@@ -31,9 +29,8 @@ class Constants {
     static final String AVRO_EXTENSION_NAME = "avro";
 
     static final String OPTION_CREATE_SETTERS = "createSetters";
-    static final String OPTION_TEMPLATE_DIR = "templateDirectory";
+    static final String OPTION_TEMPLATE_DIRECTORY = "templateDirectory";
     static final String OPTION_FIELD_VISIBILITY = "fieldVisibility";
     static final String OPTION_STRING_TYPE = "stringType";
     static final String OPTION_ENCODING = "encoding";
-
 }